fix: KEEP-1284 keep the workflow bundle free of the Solana SDK
The transaction-hash guard reached isSolanaChain through provider-factory and
validateChainTxHash through validate-chain-address. Both modules carry heavy
graphs: provider-factory pulls the RPC managers (ethers, @solana/web3.js,
safeFetch) and config-service (the db client), and validate-chain-address
imports PublicKey directly. step-success-tracker.ts is a value import of
executor.workflow.ts, so all of that landed in the workflow-function bundle,
where Node modules are rejected. The build failed with 9 esbuild errors before
next.config.ts finished loading.

Both helpers move to leaf modules that the original homes re-export, so no
caller or test mock changes its import path:

- lib/rpc/solana-chains.ts holds the chain-id set and isSolanaChain, with no
  imports at all.
- lib/web3/validate-chain-tx-hash.ts holds validateChainTxHash, which only
  ever needed a base58 decode and a length; PublicKey and ethers stay behind
  in validate-chain-address with validateChainAddress, which does need them.
  bs58 and base-x are pure ESM over Uint8Array.

Bundling step-success-tracker.ts with esbuild now resolves five modules (the
three source files plus bs58 and base-x) and references no Node builtin. The
same check against the previous imports resolves node:fs, node:async_hooks,
node:dns and node:net, which is the error set CI reported.
